# Change the function alignment of x86 to 32-bytes