--- a/tests/disas/load-store/x64/load_store_dynamic_kind_i32_index_0_guard_no_spectre_i32_access_0xffff0000_offset.wat
+++ b/tests/disas/load-store/x64/load_store_dynamic_kind_i32_index_0_guard_no_spectre_i32_access_0xffff0000_offset.wat
@@ -47,19 +47,19 @@
 ;;       movl    %edx, %r8d
 ;;       movq    %r8, %r11
 ;;       addq    0x2f(%rip), %r11
-;;       jb      0x88
-;;   67: movq    0x68(%rdi), %rsi
+;;       jb      0x98
+;;   77: movq    0x68(%rdi), %rsi
 ;;       cmpq    %rsi, %r11
-;;       ja      0x86
-;;   74: addq    0x60(%rdi), %r8
+;;       ja      0x96
+;;   84: addq    0x60(%rdi), %r8
 ;;       movl    $0xffff0000, %eax
 ;;       movl    (%r8, %rax), %eax
 ;;       movq    %rbp, %rsp
 ;;       popq    %rbp
 ;;       retq
-;;   86: ud2
-;;   88: ud2
-;;   8a: addb    %al, (%rax)
-;;   8c: addb    %al, (%rax)
-;;   8e: addb    %al, (%rax)
-;;   90: addb    $0, %al
+;;   96: ud2
+;;   98: ud2
+;;   9a: addb    %al, (%rax)
+;;   9c: addb    %al, (%rax)
+;;   9e: addb    %al, (%rax)
+;;   a0: addb    $0, %al