--- a/tests/disas/load-store/x64/load_store_dynamic_kind_i32_index_0_guard_no_spectre_i8_access_0xffff0000_offset.wat
+++ b/tests/disas/load-store/x64/load_store_dynamic_kind_i32_index_0_guard_no_spectre_i8_access_0xffff0000_offset.wat
@@ -47,21 +47,21 @@
 ;;       movl    %edx, %r8d
 ;;       movq    %r8, %r11
 ;;       addq    0x2f(%rip), %r11
-;;       jb      0x89
-;;   67: movq    0x68(%rdi), %rsi
+;;       jb      0x99
+;;   77: movq    0x68(%rdi), %rsi
 ;;       cmpq    %rsi, %r11
-;;       ja      0x87
-;;   74: addq    0x60(%rdi), %r8
+;;       ja      0x97
+;;   84: addq    0x60(%rdi), %r8
 ;;       movl    $0xffff0000, %eax
 ;;       movzbq  (%r8, %rax), %rax
 ;;       movq    %rbp, %rsp
 ;;       popq    %rbp
 ;;       retq
-;;   87: ud2
-;;   89: ud2
-;;   8b: addb    %al, (%rax)
-;;   8d: addb    %al, (%rax)
-;;   8f: addb    %al, (%rcx)
-;;   91: addb    %bh, %bh
-;;   93: incl    (%rax)
-;;   95: addb    %al, (%rax)
+;;   97: ud2
+;;   99: ud2
+;;   9b: addb    %al, (%rax)
+;;   9d: addb    %al, (%rax)
+;;   9f: addb    %al, (%rcx)
+;;   a1: addb    %bh, %bh
+;;   a3: incl    (%rax)
+;;   a5: addb    %al, (%rax)