--- a/tests/disas/load-store/x64/load_store_dynamic_kind_i64_index_0xffffffff_guard_no_spectre_i8_access_0x1000_offset.wat
+++ b/tests/disas/load-store/x64/load_store_dynamic_kind_i64_index_0xffffffff_guard_no_spectre_i8_access_0x1000_offset.wat
@@ -36,10 +36,10 @@
 ;;       movq    %rsp, %rbp
 ;;       movq    0x68(%rdi), %r8
 ;;       cmpq    %r8, %rdx
-;;       ja      0x53
-;;   41: movq    0x60(%rdi), %r10
+;;       ja      0x63
+;;   51: movq    0x60(%rdi), %r10
 ;;       movzbq  0x1000(%r10, %rdx), %rax
 ;;       movq    %rbp, %rsp
 ;;       popq    %rbp
 ;;       retq
-;;   53: ud2
+;;   63: ud2