@alexcrichton 2024-05-16
Thanks for this! It looks like this is causing changes in the disassembly of some existing tests. You can update the disassembly with `WASMTIME_TEST_BLESS=1 cargo test --test disas` locally and commit those changes to get pushed up here